third_party/WebKit/LayoutTests/fast/scrolling/scroll-into-view-small-size-ancestor.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id ="container" style="height: 300px; overflow: auto; margin-top: 200px; margin-bottom: 600px;">
    <select id="content" multiple="multiple" style="height: 600px;">
        <option>I</option>
        <option>am</option>
        <option>already</option>
        <option>in</option>
        <option>the</option>
        <option>view</option>
        <option>.</option>
        <option>Do</option>
        <option>not</option>
        <option>scroll</option>
        <option>me</option>
        <option>up</option>
        <option>.</option>
    </select>
</div>

<script>
test(function(t) {
    var content = document.getElementById("content");
    content.focus();
    assert_equals(window.scrollY, 0);
}, "Tests that when the parent is already in the view of page, the page will not scroll the parent even when the child is not completely in the view.");
</script>
